pyproject.toml
[tool.poetry]
name = "crm"
version = "0.1.0"
description = ""
authors = ["Bogdan Beliy <uxui.des@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
djangorestframework = "^3.14.0"
djoser = "^2.1.0"
pillow = "^9.4.0"
django-filter = "^22.1"
django-environ = "^0.9.0"
drf-spectacular = "^0.25.1"
psycopg2-binary = "^2.9.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
